+++ assertk/src/template/assertk/assertions/primativeArray.kt
@@ -1,5 +1,6 @@
 package assertk.assertions
 
+import kotlin.ExperimentalUnsignedTypes
 import kotlin.jvm.JvmName
 import assertk.Assert
 import assertk.all
@@ -8,25 +9,31 @@ import assertk.assertions.support.show
 import assertk.assertions.support.fail
 import assertk.assertions.support.appendName
 
-$T:$N:$E =
-    ByteArray:byteArray:Byte,
-    IntArray:intArray:Int,
-    ShortArray:shortArray:Short,
-    LongArray:longArray:Long,
-    FloatArray:floatArray:Float,
-    DoubleArray:doubleArray:Double,
-    CharArray:charArray:Char
+$T:$N:$E:$A =
+    ByteArray:byteArray:Byte:,
+    IntArray:intArray:Int:,
+    ShortArray:shortArray:Short:,
+    LongArray:longArray:Long:,
+    FloatArray:floatArray:Float:,
+    DoubleArray:doubleArray:Double:,
+    CharArray:charArray:Char:,
+    UByteArray:ubyteArray:UByte:@ExperimentalUnsignedTypes,
+    UShortArray:ushortArray:UShort:@ExperimentalUnsignedTypes,
+    UIntArray:uintArray:UInt:@ExperimentalUnsignedTypes,
+    ULongArray:ulongArray:ULong:@ExperimentalUnsignedTypes
 
 /**
  * Returns an assert on the $T's size.
  */
 @JvmName("$NSize")
+$A
 fun Assert<$T>.size() = prop("size") { it.size }
 
 /**
  * Asserts the $T contents are equal to the expected one, using [contentDeepEquals].
  * @see isNotEqualTo
  */
+$A
 fun Assert<$T>.isEqualTo(expected: $T) = given { actual ->
     if (actual.contentEquals(expected)) return
     fail(expected, actual)
@@ -36,6 +43,7 @@ fun Assert<$T>.isEqualTo(expected: $T) = given { actual ->
  * Asserts the $T contents are not equal to the expected one, using [contentDeepEquals].
  * @see isEqualTo
  */
+$A
 fun Assert<$T>.isNotEqualTo(expected: $T) = given { actual ->
     if (!(actual.contentEquals(expected))) return
     val showExpected = show(expected)
@@ -54,6 +62,7 @@ fun Assert<$T>.isNotEqualTo(expected: $T) = given { actual ->
  * @see [isNullOrEmpty]
  */
 @JvmName("$NIsEmpty")
+$A
 fun Assert<$T>.isEmpty() = given { actual ->
     if (actual.isEmpty()) return
     expected("to be empty but was:${show(actual)}")
@@ -64,6 +73,7 @@ fun Assert<$T>.isEmpty() = given { actual ->
  * @see [isEmpty]
  */
 @JvmName("$NIsNotEmpty")
+$A
 fun Assert<$T>.isNotEmpty() = given { actual ->
     if (actual.isNotEmpty()) return
     expected("to not be empty")
@@ -74,6 +84,7 @@ fun Assert<$T>.isNotEmpty() = given { actual ->
  * @see [isEmpty]
  */
 @JvmName("$NIsNullOrEmpty")
+$A
 fun Assert<$T?>.isNullOrEmpty() = given { actual ->
     if (actual == null || actual.isEmpty()) return
     expected("to be null or empty but was:${show(actual)}")
@@ -83,6 +94,7 @@ fun Assert<$T?>.isNullOrEmpty() = given { actual ->
  * Asserts the $T has the expected size.
  */
 @JvmName("$NHasSize")
+$A
 fun Assert<$T>.hasSize(size: Int) {
     size().isEqualTo(size)
 }
@@ -91,6 +103,7 @@ fun Assert<$T>.hasSize(size: Int) {
  * Asserts the $T has the same size as the expected array.
  */
 @JvmName("$NHasSameSizeAs")
+$A
 fun Assert<$T>.hasSameSizeAs(other: $T) = given { actual ->
     val actualSize = actual.size
     val otherSize = other.size
@@ -106,6 +119,7 @@ fun Assert<$T>.hasSameSizeAs(other: $T) = given { actual ->
  * ```
  */
 @JvmName("$NIndex")
+$A
 fun Assert<$T>.index(index: Int): Assert<$E> =
     transform(appendName(show(index, "[]"))) { actual ->
         if (index in 0 until actual.size) {
@@ -125,6 +139,7 @@ fun Assert<$T>.index(index: Int): Assert<$E> =
  * ```
  */
 @JvmName("$NEach")
+$A
 fun Assert<$T>.each(f: (Assert<$E>) -> Unit) = given { actual ->
     all {
         actual.forEachIndexed { index, item ->